Move lookup plugin directory. Create config loader test
